requirements.txt
Flask-Caching==2.1.0
flask-cors
python-dotenv
matplotlib
numpy
langchain-openai
openai
langchain